backend/requirements.txt
# FastAPI 和服务器
fastapi==0.115.6
uvicorn[standard]==0.34.0
websockets==14.1
python-multipart==0.0.20

# Telegram MTProto
Telethon==1.40.0

# 数据库
sqlalchemy==2.0.36
alembic==1.14.0
aiomysql==0.2.0

# 认证和安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# HTTP 客户端
httpx==0.28.1
aiohttp==3.11.11

# YAML 解析（代理配置需要）
pyyaml==6.0.1

# 通知推送
requests==2.32.3
pydantic==2.10.4
pydantic-settings==2.7.1

# 代理支持
pysocks==1.7.1

# 词云
wordcloud==1.9.4
jieba==0.42.1

# 其他工具
loguru==0.7.3
apscheduler==3.11.0
python-dateutil==2.9.0.post0
psutil==6.1.1  # 系统资源监控
